Add keyboard HID interface to Solo
This patch adds a keyboard HID interface to the existing FIDO2 HID
interface. It allows injecting keystokes into the host when the Solo
button is pressed. For now it injects the hardcoded sequence "1234".

The new interface is enabled with the ENABLE_KBD macro. Right now it
doesn't work with the other additional interfaces (CCID and CDC).
Advertising a random combination of USB interfaces (FIDO2, kbd, CCID,
CDC) requires more work and could be addressed in the future.

Issue: #354